feat: Add .gitignore file
Instantiated .gitignore with sections for editor-specific directories,
python virtual environment directories, and the `ConvertMe` directory
used to compile .frag files to .qsb.

The directories for virtual environments are handy if the user uses
`conda` or `venv` to manage python packages but needs to symlink to a
directory within the root of the repo to ensure that packages are
readable/executable with their IDE/Editor of choice within the
scope of the project.